Specify the type of certificate to import
Certificates can be exported in Windows using 4 different formats:
PKCS12, PKCS7, DER encoded X509 and Base64 encoded X509. The
Import-PfxCertificate only supports PKCS12/PFX, for all other types it
is necessary to use the Import-Certificate method instead.

I added an additional parameter to the CertificateStore specifying the
format of the certificate, allowing the user to choose from PKCS12
(default), PKCS7 and X509 (both DER and Base64), and implemented a check
to determine which method to use while importing.